# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Household task management API.\n\nConventions callers need:\n- Every response is the envelope `{ success, message, data }`.\n- `user`, `group` and `task` are addressed by opaque `uid`; every other resource by numeric `id`.\n- Timestamps are UTC with no zone suffix — parse as UTC, convert for display.\n- Actors are derived from the token; never send a user or member id to identify the caller.\n- Errors: 400 validation, 401 unauthenticated, 403 forbidden, 404 not found, 409 duplicate or violated business rule, 429 rate limited (see `Retry-After`).\n- Role requirements are stated per operation and are not modelled in this spec.\n",
    "title" : "TasKlean API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Relative to the host serving this API"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/audit-logs/member/{groupMemberId}" : {
      "get" : {
        "description" : "Lists audit-log entries recorded for a specific group member, newest first.",
        "operationId" : "getLogsByMember",
        "parameters" : [ {
          "description" : "the group member id",
          "in" : "path",
          "name" : "groupMemberId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAuditLogResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the entries"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Lists audit-log entries recorded for a specific group member, newest first.",
        "tags" : [ "Audit logs" ]
      }
    },
...
    "/api/group-members" : {
      "get" : {
        "description" : "Lists a group's active members.",
        "operationId" : "getMembersByGroup",
        "parameters" : [ {
          "description" : "the group id",
          "in" : "query",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListGroupMemberResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the members"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Lists a group's active members.",
        "tags" : [ "Group members" ]
      },
      "post" : {
        "description" : "Adds a user to a group.",
        "operationId" : "addMember",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GroupMemberRequest"
              }
            }
          },
          "description" : "the membership details (user, group, role)",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGroupMemberResponse"
                }
              }
            },
            "description" : "<code>201 Created</code> with the created membership"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Adds a user to a group.",
        "tags" : [ "Group members" ]
      }
    },
    "/api/group-members/{id}" : {
      "delete" : {
        "description" : "Soft-removes a member from a group.",
        "operationId" : "removeMember",
        "parameters" : [ {
          "description" : "the membership id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "<code>200 OK</code> with a confirmation message"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Soft-removes a member from a group.",
        "tags" : [ "Group members" ]
      },
      "get" : {
        "description" : "Fetches a membership by id.",
        "operationId" : "getMember",
        "parameters" : [ {
          "description" : "the membership id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGroupMemberResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the membership"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Fetches a membership by id.",
        "tags" : [ "Group members" ]
      }
    },
    "/api/group-members/{id}/role" : {
      "patch" : {
        "description" : "Changes a member's role.",
        "operationId" : "updateRole",
        "parameters" : [ {
          "description" : "the membership id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "description" : "the new role",
          "in" : "query",
          "name" : "role",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "GROUP_ADMIN", "GROUP_MEMBER" ]
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGroupMemberResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the updated membership"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Changes a member's role.",
        "tags" : [ "Group members" ]
      }
    },
...
    "/api/groups/{uid}" : {
      "delete" : {
        "description" : "Soft-deletes a group.",
        "operationId" : "deleteGroup",
        "parameters" : [ {
          "description" : "the group's public UID",
          "in" : "path",
          "name" : "uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "<code>200 OK</code> with a confirmation message"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Soft-deletes a group.",
        "tags" : [ "Groups" ]
      },
      "get" : {
        "description" : "Fetches a group by public UID.",
        "operationId" : "getGroup",
        "parameters" : [ {
          "description" : "the group's public UID",
          "in" : "path",
          "name" : "uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGroupResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the group"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Fetches a group by public UID.",
        "tags" : [ "Groups" ]
      },
      "put" : {
        "description" : "Updates a group.",
        "operationId" : "updateGroup",
        "parameters" : [ {
          "description" : "the group's public UID",
          "in" : "path",
          "name" : "uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GroupRequest"
              }
            }
          },
          "description" : "the new group details",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGroupResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the updated group"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Updates a group.",
        "tags" : [ "Groups" ]
      }
    },
...
    "/api/task-completions/member/{groupMemberId}" : {
      "get" : {
        "description" : "Lists the completions recorded by a group member, newest first.",
        "operationId" : "getCompletionsByMember",
        "parameters" : [ {
          "description" : "the group member id",
          "in" : "path",
          "name" : "groupMemberId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTaskCompletionResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the completions"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Lists the completions recorded by a group member, newest first.",
        "tags" : [ "Task completions" ]
      }
    },
...
  },
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Validation failure, malformed body, or a bad path/query parameter"
      },
      "Forbidden" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Authenticated, but lacking the platform or group role this operation requires"
      },
      "InternalServerError" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Unexpected server error"
      },
      "TooManyRequests" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Rate limit exceeded — see the Retry-After header"
      },
      "Unauthorized" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "No valid access token — or, on the public auth endpoints, rejected credentials"
      }
    },
    "schemas" : {
...
      "ApiResponseGroupMemberResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/GroupMemberResponse"
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "ApiResponseGroupResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/GroupResponse"
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "ApiResponseListAuditLogResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditLogResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseListGroupMemberResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GroupMemberResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseListTaskCompletionResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TaskCompletionResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseVoid" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : { },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "AuditLogResponse" : {
        "type" : "object",
        "properties" : {
          "action" : {
            "type" : "string"
          },
          "actionMessage" : {
            "type" : "string"
          },
          "actorUserId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "entityId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "entityType" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "groupMemberId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ipAddress" : {
            "type" : "string"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "description" : "The standard envelope, as returned for a failed request.",
        "properties" : {
          "data" : {
            "type" : "object",
            "description" : "Always null on an error"
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message"
          },
          "success" : {
            "type" : "boolean",
            "default" : false
          }
        }
      },
...
      "GroupMemberRequest" : {
        "type" : "object",
        "properties" : {
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "GROUP_ADMIN", "GROUP_MEMBER" ]
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "groupId", "role", "userId" ]
      },
      "GroupMemberResponse" : {
        "type" : "object",
        "properties" : {
          "addedById" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dateJoined" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateLeft" : {
            "type" : "string",
            "format" : "date-time"
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "removedById" : {
            "type" : "integer",
            "format" : "int64"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "GROUP_ADMIN", "GROUP_MEMBER" ]
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "GroupRequest" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "photoUrl" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ]
      },
      "GroupResponse" : {
        "type" : "object",
        "description" : "A group as returned by the API. The invite code is a shareable join secret, so it is exposed\n only through {@link com.tasklean.api.domain.group.dto.GroupResponse#withInviteCode(Group) #withInviteCode(Group)} — for callers allowed to invite.",
        "properties" : {
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateUpdated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "inviteCode" : {
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "photoUrl" : {
            "type" : "string"
          },
          "uid" : {
            "type" : "string"
          }
        }
      },
...
      "TaskCompletionResponse" : {
        "type" : "object",
        "properties" : {
          "completionNote" : {
            "type" : "string"
          },
          "completionPhotoUrl" : {
            "type" : "string"
          },
          "dateCompleted" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "groupMemberId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "taskId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from /api/auth/login, /api/auth/verify-email, /api/auth/google or /api/auth/refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}